Issue #68: Lessen dependency on Guava
Reduces Guava dependency as discussed in Issue #68.

With these changes the only Guava dependencies left in the source of
the plugin are VisibleForTesting. As this is only an annotation for
documentation purposes it is inconsequential if it can't be loaded at
runtime. This annotation can not be removed due to existing PMD rules
preventing package scoping.

The tests still depend on Guava but this is fine.